refactor(main): register IPC per domain, and move guidance next to what it guards
`src/main/index.ts` held 128 endpoint registrations inline, so every main-process
change touched the same 2,563-line file. Each domain now owns a module under
`src/main/ipc/`, matching the `register-team-handlers.ts` that was already there,
and `registerIpcHandlers` is a 16-call dispatcher over a named
`IpcHandlerDependencies` object. index.ts drops to 1,620 lines.

No behaviour changes. Every registration moved verbatim except seven: four
handlers gain `const mainWindow = getMainWindow();` and two call the accessor
inline, because a registrar receives an accessor where index.ts closed over a
module-scope binding, and one calls the injected
`setAnalyticsTrackingEnabled` instead of reaching for `hostAnalytics` directly.
The payload decoders `#195` made mandatory are carried across untouched.

A module the dispatcher never calls would otherwise be invisible: the coverage
scan reads sources, so an orphaned registrar still looks registered while every
channel in it rejects at runtime, and `tsc` catches only half of that — deleting
the call alone is `TS6133`, but deleting the import with it compiles, and so does
a registrar nobody wired up. `ipc-channel-coverage.test.ts` gains "calls every
registrar under src/main/ipc exactly once from the dispatcher" to close the rest.

The root `AGENTS.md` was 279 lines and growing, most of it rules that only apply
inside one directory. Five directories now carry their own, loaded when a file
under them is opened: the renderer stack, this trust boundary, the user's SQLite,
the Team API wire protocol, and the account Worker's D1 migrations — that last
one moved out of `src/backend/AGENTS.md`, which is not an ancestor of
`apps/auth-api/migrations/` and so was never loaded where it applied. Root drops
to 202 lines and keeps what is repository-wide.

Surfaces touched: desktop main process and repository documentation only — no
renderer, mobile, web, palette, IPC contract, migration or reverse-state change.

Tests: watched the new coverage assertion fail by deleting both the
`registerMemoryIpcHandlers` call and its import, which `bun run typecheck:node`
accepts at exit 0 and the assertion rejects by name.

Re-applies #191, reverted in #197, on top of #195.
